ERB likes this better. i was actually testing against erubis...
not sure if there's a way to enforce using erb. we can't set it
in test.rb since it's called in logi-list, so we need to set there.
however logi-list cannot tell if we're testing or not, and i also
don't want to make this an enforcement.

i guess i'll leave it is for now, and let travis use bundler to
solve this issue. personally i don't want to use bundler for my
runtime, which would pollute the environment